checking whether averaging climatic variables over the estimated period of time overcomes the local minima issues
